move PokemonCard file to containers folder and move app and routes files to components folder
